fix testing and some better practises.
Testing just checks game no longer calls another game.Gamestart() as this was not needed.

Testing also was in a while loop likely as a result of copying from program however this has been changed.

further more the game function was access the private sum directly this has been chnaged.